requirements.txt
flask==2.3.3
httpx==0.27.0
python-dotenv==1.0.1
